requirements.txt
# No external dependencies required for heap sort
# Python built-in modules used: sys, json, random

fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
docker==7.0.0
google-adk==1.0.0
supabase==1.2.1
dotenv==1.0.1